refactor(davinci): add shared physical stage identifiers
